# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/recipes": {
            "get": {
                "description": "Get a list of all available recipes.",
                "produces": [
                    "application/json"
                ],
                "summary": "Get all recipes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/pkg.Recipe"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new recipe.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create a new recipe",
                "parameters": [
                    {
                        "description": "Recipe object to be created",
                        "name": "recipe",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg.Recipe"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.Recipe"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/recipes/search": {
            "get": {
                "description": "Search for recipes that contain a specific tag.",
                "produces": [
                    "application/json"
                ],
                "summary": "Search recipes by tag",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tag to search for",
                        "name": "tag",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/pkg.Recipe"
                            }
                        }
                    },
                    "404": {
                        "description": "No recipes found with that tag",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/recipes/{id}": {
            "get": {
                "description": "Return a recipe that contains a specific id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Return recipes by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the recipe to list",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Recipe object to be updated",
                        "name": "recipe",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg.Recipe"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.Recipe"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Recipe not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update a recipe that contains a specific id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update recipes by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the recipe to update",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Recipe object to be updated",
                        "name": "recipe",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg.Recipe"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.Recipe"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Recipe not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a recipe that contains a specific id.",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete recipes by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the recipe to delete",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Recipe not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/refresh": {
            "post": {
                "description": "Refreshes an expired JWT token.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh JWT token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token refreshed successfully",
                        "schema": {
                            "$ref": "#/definitions/auth.jWTOutput"
                        }
                    },
                    "400": {
                        "description": "Token is not expired yet",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to generate token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/signin": {
            "post": {
                "description": "Authenticates a user and returns a JWT token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Sign in a user",
                "parameters": [
                    {
                        "description": "User credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.user"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully signed in",
                        "schema": {
                            "$ref": "#/definitions/auth.jWTOutput"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid username or password",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to generate token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "auth.jWTOutput": {
            "type": "object",
            "properties": {
                "expires": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "auth.user": {
            "type": "object",
            "properties": {
                "password": {
                    "description": "User's password.",
                    "type": "string"
                },
                "username": {
                    "description": "User's username.",
                    "type": "string"
                }
            }
        },
        "pkg.Recipe": {
            "description": "Represents a culinary recipe with a name, ingredients, and tags.",
            "type": "object",
            "properties": {
                "id": {
                    "description": "ID of the recipe.\n@Example 123ojwfnowndno",
                    "type": "string"
                },
                "ingredients": {
                    "description": "List of ingredients required for the recipe.\n@Example [\"pasta\", \"tomato sauce\", \"cheese\"]",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "instructions": {
                    "description": "Instructions associated with the recipe.\n@Example [\"To marinate the chicken\", \"Add scallion whites, and cook, stirring\"]",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "description": "Name of the recipe.\n@Example Delicious Pasta",
                    "type": "string"
                },
                "publishedAt": {
                    "description": "Time of  the recipe added or updated .\n@Example \"2021-01-17T19:28:52.803062+01:00\"",
                    "type": "string"
                },
                "tags": {
                    "description": "Tags associated with the recipe.\n@Example [\"italian\", \"dinner\"]",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}